requirements.txt
Django==4.1.2
django-phonenumber-field[phonenumberslite]
django-countries
crispy-bootstrap5